requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
mysql-connector-python>=8.0.30
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0